requirements.txt
argcomplete==1.9.4
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.1.1
PyYAML>=3.13
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
